feat: move runner impl for client into extra module; impl procedure execution into Client
