+++ client_new/src/procedures.rs
@@ -1,8 +1,11 @@
 // Copyright 2020-2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
+mod clientrunner;
 mod primitives;
 mod types;
+
+pub use clientrunner::*;
 
 pub use primitives::{
     AeadCipher, AeadDecrypt, AeadEncrypt, BIP39Generate, BIP39Recover, Chain, ChainCode, CopyRecord, Ed25519Sign,
